requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
pydub>=0.25.0

# Evaluation Metrics
pesq>=0.0.4
pystoi>=0.3.3
mir_eval>=0.7

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.2.0
wandb>=0.15.0  # Optional: for experiment tracking

# Web API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pytest>=7.3.0
